refactor: rename `Resizable` to `Moveable`
