+++ src/routes/[id]/components/moveable.svelte
