+++ src/routes/[id]/components/trimmer.svelte
@@ -2,7 +2,7 @@
 	import { edits } from '../stores/edits.store';
 	import { recording } from '../stores/recording.store';
 	import { videoStatus } from '../stores/video-status.store';
-	import Resizable from './resizable.svelte';
+	import Moveable from './moveable.svelte';
 
 	export let isResizing: boolean;
 	const MIN_VIDEO_DURATION_IN_SECONDS = 1;
@@ -10,7 +10,7 @@
 	$: left = ($edits.startAt * 100) / $recording?.duration;
 </script>
 
-<Resizable
+<Moveable
 	className={{
 		root: `group h-10 bg-white/5 border-2 border-white/10 rounded-lg absolute ring ring-transparent ring-offset-0 [&.current-trim]:bg-blue-700/20 [&.current-trim]:border-blue-700/50 [&.current-trim]:focus-within:ring-blue-700/20 hover:bg-blue-700/20 hover:border-blue-700/50 has-[:active]:bg-blue-700/20 has-[:active]:border-blue-700/50 focus-within:ring-white/5 focus-within:hover:ring-blue-700/20 hover:z-10 ${
 			$videoStatus.currentTime >= $edits.startAt && $videoStatus.currentTime <= $edits.endAt
@@ -53,4 +53,4 @@
 	<div slot="e" class="w-[12px] h-[75%] bg-blue-900 rounded-r-md flex justify-center items-center">
 		<div class="w-[2px] h-[45%] bg-neutral-50/50 rounded-full" />
 	</div>
-</Resizable>
+</Moveable>
